fix(android): destroy WebViews on shutdown, fix mDNS re-register crash, acquire MulticastLock, drop dead WebViewContainer
